Fixed task 2 and 3

+++ 2_DataTypes_Variables_Operators_Memory/src/Task3_CalculateExpression.java
@@ -3,7 +3,7 @@ public class Task3_CalculateExpression
 {
 	public static void main(String args[])
 	{
-		double result = ( 3291 + 88581 ) / 14 + ( 1116 % 171 ) * 5 - 312 / ( 4 + 18 );
+		int result = ( 3291 + 88581 ) / 14 + ( 1116 % 171 ) * 5 - 312 / ( 4 + 18 );
 		System.out.println(result);
 	}
 }
